+++ src/components/mod.rs
@@ -35,6 +35,7 @@ pub mod project_manager;
 pub mod report_issue_button;
 pub mod settings;
 pub mod routing_rule_editor;
+pub mod station_label_tooltip;
 pub mod tab_view;
 pub mod text_input_dialog;
 pub mod time_graph;